+++ store/primary/src/store/actions.rs
@@ -1,5 +1,4 @@
 use opentracingrust::SpanContext;
-use uuid::Uuid;
 
 use replicante_models_core::actions::node::ActionSyncSummary;
 
@@ -16,22 +15,6 @@ pub struct Actions {
 impl Actions {
     pub(crate) fn new(actions: ActionsImpl, attrs: ActionsAttributes) -> Actions {
         Actions { actions, attrs }
-    }
-
-    /// Approve a PENDING_APPROVE action for scheduling.
-    pub fn approve<S>(&self, action_id: Uuid, span: S) -> Result<()>
-    where
-        S: Into<Option<SpanContext>>,
-    {
-        self.actions.approve(&self.attrs, action_id, span.into())
-    }
-
-    /// Disapprove a PENDING_APPROVE action so it won't be scheduled.
-    pub fn disapprove<S>(&self, action_id: Uuid, span: S) -> Result<()>
-    where
-        S: Into<Option<SpanContext>>,
-    {
-        self.actions.disapprove(&self.attrs, action_id, span.into())
     }
 
     /// Iterate all unfinished actions for the cluster returning only summary information.